requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
astroid==2.9.3
black==22.1.0
certifi==2021.10.8
cloudinary==1.29.0
colorama==0.4.4
Deprecated==1.2.13
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.2
django-appconf==1.0.5
django-cacheops==6.0
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-heroku==0.3.1
django-imagekit==4.1.0
django-ipware==4.0.2
django-pickling==1.0
django-polymorphic==3.1.0
django-rest-passwordreset==1.2.1
django-rest-polymorphic==0.1.9
django-reversion==5.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.4.2
funcy==1.17
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.22.3
packaging==21.3
pathspec==0.9.0
pilkit==2.0
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==2.5.0
promise==2.3
psycopg2==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
python-http-client==3.3.6
pytz==2021.3
redis==4.1.4
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.1
virtualenv-clone==0.5.7
whitenoise==6.0.0
wrapt==1.13.3
